How to make Tick, Tak Toe in the Python(Single/multiplayer)
Using the Al in my code and making a Tick, Tak Toe in the language of the python. Making a single player and the code enables not the user to win the game but the game can be draw. And you can also make the Single player and the Multiplayer game.